little refactoring in imports
