+++ src/common/interfaces/user-response.interface.ts
@@ -1,4 +1,4 @@
-import { FavoriteProductsInterface } from '../models/favorite-products.model';
+import { FavoriteProductsInterface } from '../models/product.favorite-products.model';
 import { UserToOrdersInterface } from '../models/order.userToOrders.model';
 import { UserType } from '../types/user.type';
